+++ include/dataStructures/Pin.h
@@ -65,11 +65,11 @@ class Pin {
   // minExpSum_: holds exp(-x_i/gamma)
   // the x_i is equal to cx_ variable.
   //
-  float maxExpSumX_{};
-  float maxExpSumY_{};
+  double maxExpSumX_{};
+  double maxExpSumY_{};
 
-  float minExpSumX_{};
-  float minExpSumY_{};
+  double minExpSumX_{};
+  double minExpSumY_{};
 
   // flag variables
   //
@@ -146,10 +146,10 @@ class Pin {
   void setMaxExpSumY(float maxExpSumY);
   void setMinExpSumX(float minExpSumX);
   void setMinExpSumY(float minExpSumY);
-  float maxExpSumX() const { return maxExpSumX_; }
-  float maxExpSumY() const { return maxExpSumY_; }
-  float minExpSumX() const { return minExpSumX_; }
-  float minExpSumY() const { return minExpSumY_; }
+  double maxExpSumX() const { return maxExpSumX_; }
+  double maxExpSumY() const { return maxExpSumY_; }
+  double minExpSumX() const { return minExpSumX_; }
+  double minExpSumY() const { return minExpSumY_; }
 
   bool hasMaxExpSumX() const { return (hasMaxExpSumX_ == 1); }
   bool hasMaxExpSumY() const { return (hasMaxExpSumY_ == 1); }
